+++ lms-api/src/main/java/com/savelms/api/vacation/dto/UseVacationRequest.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.NotNull;
 public class UseVacationRequest {
 
     @NotNull
-    private Long usedDays;
+    private Double usedDays;
 
     @NotNull
     private String reason;
